+++ public/js/ajax/game/ajax.js
@@ -59,29 +59,31 @@ function removeGameFromFavourite(e,gameID){
 function deleteGame(e,gameID){
   if(e!=null) e.preventDefault();
 
-  $.ajax({
-    type: "DELETE",
-    url: '/game/'+gameID,
-    dataType: 'json',
-    data: "_token="+$("[name='_token']").val(),
-    success: function(data){
-      if(data.success != undefined){
-				$('#row-game-'+gameID).html('');
-        $('#response-success').show();
-        $('#response-error').hide();
-        $('#response-success').html(data.success);
-      }else{
-        $('#response-error').show();
-        $('#response-success').hide();
-        $('#response-error').html(data.error);
-      }
-    },
-    error: function(xhr,status,error){
-      $('#response-error').show();
-      $('#response-success').hide();
-      $('#response-error').html("An error occured on the server, please try again later.");
-    }
-  });
+	if(confirm("Do you really want to delete this game ?")){
+	  $.ajax({
+	    type: "DELETE",
+	    url: '/game/'+gameID,
+	    dataType: 'json',
+	    data: "_token="+$("[name='_token']").val(),
+	    success: function(data){
+	      if(data.success != undefined){
+					$('#row-game-'+gameID).html('');
+	        $('#response-success').show();
+	        $('#response-error').hide();
+	        $('#response-success').html(data.success);
+	      }else{
+	        $('#response-error').show();
+	        $('#response-success').hide();
+	        $('#response-error').html(data.error);
+	      }
+	    },
+	    error: function(xhr,status,error){
+	      $('#response-error').show();
+	      $('#response-success').hide();
+	      $('#response-error').html("An error occured on the server, please try again later.");
+	    }
+	  });
+	}
 
   return false;
 }
